requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
async-timeout==4.0.3
    # via redis
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
distro==1.9.0
    # via openai
fastapi==0.110.1
    # via -r requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
openai==1.21.0
    # via -r requirements/requirements.in
pydantic==2.7.0
    # via
    #   fastapi
    #   openai
pydantic-core==2.18.1
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements/requirements.in
redis==5.0.3
    # via -r requirements/requirements.in
requests==2.31.0
    # via -r requirements/requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
uvicorn==0.29.0
    # via -r requirements/requirements.in
